making black happy

+++ orchestration/utils/data_preparation/clean.py
@@ -4,11 +4,33 @@ from prefect import task
 
 def clean(df: pd.DataFrame) -> pd.DataFrame:
     # Renaming columns
-    df = df.rename(columns={'oldbalanceOrg': 'oldbalanceOrig'})
+    df = df.rename(columns={"oldbalanceOrg": "oldbalanceOrig"})
 
     # Transform type to category to optimize memory and speed
-    df['step'] = df['step'].astype('int16')
-    df[['type', 'nameOrig', 'nameDest']] = df[['type', 'nameOrig', 'nameDest']].astype('category')
-    df[['amount', 'oldbalanceOrig', 'newbalanceOrig', 'oldbalanceDest', 'newbalanceDest', 'isFraud']] = df[['amount', 'oldbalanceOrig', 'newbalanceOrig', 'oldbalanceDest', 'newbalanceDest', 'isFraud']].astype('float32')
+    df["step"] = df["step"].astype("int16")
+    df[["type", "nameOrig", "nameDest"]] = df[["type", "nameOrig", "nameDest"]].astype(
+        "category"
+    )
+    df[
+        [
+            "amount",
+            "oldbalanceOrig",
+            "newbalanceOrig",
+            "oldbalanceDest",
+            "newbalanceDest",
+            "isFraud",
+        ]
+    ] = df[
+        [
+            "amount",
+            "oldbalanceOrig",
+            "newbalanceOrig",
+            "oldbalanceDest",
+            "newbalanceDest",
+            "isFraud",
+        ]
+    ].astype(
+        "float32"
+    )
 
     return df
